+++ seiscat/utils/__init__.py
@@ -10,3 +10,4 @@ Utility functions for seiscat.
     (https://www.gnu.org/licenses/gpl-3.0-standalone.html)
 """
 from .exit import err_exit, ExceptionExit  # noqa
+from .conversion import float_or_none, int_or_none  # noqa